Better aspect ratio implementation

+++ Source/render/TextRenderer.swift
@@ -145,7 +145,7 @@ class TextRenderer: NodeRenderer {
             NSAttributedStringKey.font: font
         ]
         let textSize = NSString(string: text.text).size(withAttributes: textAttributes)
-        return -text.align.align(x: textSize.width)
+        return -CGFloat(text.align.align(x: textSize.width.doubleValue))
     }
 
     fileprivate func getTextColor(_ fill: Fill) -> MColor {
